StudyHub/pdfbackend/fastapibackend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bibtexparser==2.0.0b6
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.22.1
colorama==0.4.6
colorlog==6.8.2
comm==0.2.1
cryptography==42.0.2
csvw==3.3.0
Cython==3.0.8
dataclasses-json==0.6.4
debugpy==1.8.0
decorator==5.1.1
dlinfo==1.2.1
executing==2.0.1
faiss-cpu==1.7.4
filelock==3.13.1
Flask==3.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
frozenlist==1.4.1
fsspec==2023.12.2
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.5.1
gunicorn==21.2.0
idna==3.6
ipykernel==6.29.0
ipython==8.21.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langchain-google-genai==0.0.6
langsmith==0.0.86
language-tags==1.2.0
lazy_loader==0.3
libretranslatepy==2.1.1
librosa==0.10.1
llvmlite==0.41.1
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib-inline==0.1.6
monotonic-align==1.0.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
packaging==23.2
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.3
phonemizer==3.2.1
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydub==0.25.1
Pygments==2.17.2
pylatexenc==2.10
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.26.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rdflib==7.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
segments==2.2.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.1
SQLAlchemy==2.0.25
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
translate==3.6.1
ttsmms==0.7
typing-inspect==0.9.0
typing_extensions==4.9.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.0
uuid==1.30
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.9.4
